refactor: Extract `onKeyDown` handler hook from `Form` component
